Renamed notifier module
